pyproject.toml
[project]
name = "google-maps-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "uvicorn[standard]>=0.34.2",
]
